chore: rename schema file
