+++ packages/netlify-cms-widget-code/src/schema.js
