+++ packages/netlify-cms-widget-datetime/src/schema.js
