+++ packages/netlify-cms-widget-file/src/schema.js
