+++ packages/netlify-cms-widget-image/src/index.js
@@ -1,6 +1,6 @@
 import NetlifyCmsWidgetFile from 'netlify-cms-widget-file';
 import previewComponent from './ImagePreview';
-import schema from './Schema';
+import schema from './schema';
 
 const controlComponent = NetlifyCmsWidgetFile.withFileControl({ forImage: true });
 const Widget = (opts = {}) => ({
