+++ packages/netlify-cms-widget-image/src/schema.js
