+++ packages/netlify-cms-widget-list/src/schema.js
