+++ packages/netlify-cms-widget-map/src/schema.js
