+++ packages/netlify-cms-widget-markdown/src/index.js
@@ -1,6 +1,6 @@
 import controlComponent from './MarkdownControl';
 import previewComponent from './MarkdownPreview';
-import schema from './Schema';
+import schema from './schema';
 
 const Widget = (opts = {}) => ({
   name: 'markdown',
