+++ packages/netlify-cms-widget-markdown/src/schema.js
