+++ packages/netlify-cms-widget-number/src/schema.js
