+++ packages/netlify-cms-widget-object/src/schema.js
