+++ packages/netlify-cms-widget-relation/src/index.js
@@ -1,6 +1,6 @@
 import controlComponent from './RelationControl';
 import previewComponent from './RelationPreview';
-import schema from './Schema';
+import schema from './schema';
 
 const Widget = (opts = {}) => ({
   name: 'relation',
