+++ packages/netlify-cms-widget-relation/src/schema.js
