+++ packages/netlify-cms-widget-select/src/index.js
@@ -1,6 +1,6 @@
 import controlComponent from './SelectControl';
 import previewComponent from './SelectPreview';
-import schema from './Schema';
+import schema from './schema';
 
 const Widget = (opts = {}) => ({
   name: 'select',
