+++ packages/netlify-cms-widget-select/src/schema.js
